+++ src/test/java/nextstep/subway/unit/PathFinderImplTest.java
@@ -1,9 +1,11 @@
 package nextstep.subway.unit;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 import java.util.List;
+import nextstep.common.exception.CustomException;
 import nextstep.subway.domain.Line;
 import nextstep.subway.domain.Station;
 import nextstep.subway.domain.path.Path;
@@ -14,7 +16,7 @@ import org.assertj.core.util.Lists;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-public class PathDistanceFinderTest {
+public class PathFinderImplTest {
 
   private Station 교대역;
   private Station 강남역;
@@ -75,5 +77,51 @@ public class PathDistanceFinderTest {
         () -> assertThat(path.extractDuration()).isEqualTo(9),
         () -> assertThat(path.extractFare()).isEqualTo(1_250)
     );
+  }
+
+  @Test
+  void 최소_시간_경로_조회() {
+    // given
+    List<Line> lines = Lists.newArrayList(신분당선, 이호선, 삼호선);
+    PathFinder pathFinder = new PathFinderImpl(lines);
+
+    // when
+    Path path = pathFinder.findPath(교대역, 양재역, PathType.DURATION);
+
+    // then
+    assertAll(
+        () -> assertThat(path.getStations()).containsExactlyInAnyOrder(교대역, 강남역, 양재역),
+        () -> assertThat(path.extractDistance()).isEqualTo(9),
+        () -> assertThat(path.extractDuration()).isEqualTo(9),
+        () -> assertThat(path.extractFare()).isEqualTo(1_250)
+    );
+  }
+
+  @Test
+  void 반대로_최소_시간_경로_조회() {
+    // given
+    List<Line> lines = Lists.newArrayList(신분당선, 이호선, 삼호선);
+    PathFinder pathFinder = new PathFinderImpl(lines);
+
+    // when
+    Path path = pathFinder.findPath(양재역, 교대역, PathType.DURATION);
+
+    // then
+    assertAll(
+        () -> assertThat(path.getStations()).containsExactlyInAnyOrder(양재역, 강남역, 교대역),
+        () -> assertThat(path.extractDistance()).isEqualTo(9),
+        () -> assertThat(path.extractDuration()).isEqualTo(9),
+        () -> assertThat(path.extractFare()).isEqualTo(1_250)
+    );
+  }
+
+  @Test
+  void 경로_조회_타입_에러() {
+    // given
+    List<Line> lines = Lists.newArrayList(신분당선, 이호선, 삼호선);
+    PathFinder pathFinder = new PathFinderImpl(lines);
+
+    // when
+    assertThatThrownBy(() -> pathFinder.findPath(양재역, 교대역, null)).isInstanceOf(CustomException.class);
   }
 }
