+++ src/test/java/nextstep/subway/unit/PathServiceTest.java
@@ -70,24 +70,14 @@ public class PathServiceTest {
   void 두_역_최단_거리_조회() {
     PathResponse result = pathService.findPath(교대역, 양재역, PathType.DISTANCE);
 
-    assertAll(
-        () -> assertThat(result.getStations().size()).isEqualTo(3),
-        () -> assertThat(result.getDistance()).isEqualTo(6),
-        () -> assertThat(result.getDuration()).isEqualTo(5),
-        () -> assertThat(result.getFare()).isEqualTo(1_250)
-    );
+    경로_조회_정보_비교(result, 3, 6, 5, 1_250);
   }
 
   @Test
   void 두_역_최소_시간_조회() {
     PathResponse result = pathService.findPath(교대역, 양재역, PathType.DURATION);
 
-    assertAll(
-        () -> assertThat(result.getStations().size()).isEqualTo(3),
-        () -> assertThat(result.getDistance()).isEqualTo(6),
-        () -> assertThat(result.getDuration()).isEqualTo(5),
-        () -> assertThat(result.getFare()).isEqualTo(1_250)
-    );
+    경로_조회_정보_비교(result, 3, 6, 5, 1_250);
   }
 
   @Test
@@ -99,71 +89,50 @@ public class PathServiceTest {
   void 두_역_10KM_이내_최단_거리_경로_조회() {
     PathResponse result = pathService.findPath(교대역, 양재역, PathType.DISTANCE);
 
-    assertAll(
-        () -> assertThat(result.getStations().size()).isEqualTo(3),
-        () -> assertThat(result.getDistance()).isEqualTo(6),
-        () -> assertThat(result.getDuration()).isEqualTo(5),
-        () -> assertThat(result.getFare()).isEqualTo(1_250)
-    );
+    경로_조회_정보_비교(result, 3, 6, 5, 1_250);
   }
 
   @Test
   void 두_역_10KM_초과_50KM_이하_최단_거리_경로_조회() {
     PathResponse result = pathService.findPath(남부터미널역, 강남역, PathType.DISTANCE);
 
-    assertAll(
-        () -> assertThat(result.getStations().size()).isEqualTo(3),
-        () -> assertThat(result.getDistance()).isEqualTo(18),
-        () -> assertThat(result.getDuration()).isEqualTo(6),
-        () -> assertThat(result.getFare()).isEqualTo(1_450)
-    );
+    경로_조회_정보_비교(result, 3, 18, 6, 1_450);
   }
 
   @Test
   void 두_역_경로_50KM_초과_최단_거리_경로_조회() {
     PathResponse result = pathService.findPath(양재역, 대치역, PathType.DISTANCE);
 
-    assertAll(
-        () -> assertThat(result.getStations().size()).isEqualTo(3),
-        () -> assertThat(result.getDistance()).isEqualTo(56),
-        () -> assertThat(result.getDuration()).isEqualTo(8),
-    () -> assertThat(result.getFare()).isEqualTo(2_150)
-    );
+    경로_조회_정보_비교(result, 3, 56, 8, 2_150);
   }
 
   @Test
   void 두_역_10KM_이내_최소_시간_경로_조회() {
     PathResponse result = pathService.findPath(교대역, 양재역, PathType.DURATION);
 
-    assertAll(
-        () -> assertThat(result.getStations().size()).isEqualTo(3),
-        () -> assertThat(result.getDistance()).isEqualTo(6),
-        () -> assertThat(result.getDuration()).isEqualTo(5),
-        () -> assertThat(result.getFare()).isEqualTo(1_250)
-    );
+    경로_조회_정보_비교(result, 3, 6, 5, 1_250);
   }
 
   @Test
   void 두_역_10KM_초과_50KM_이하_최소_시간_경로_조회() {
     PathResponse result = pathService.findPath(남부터미널역, 강남역, PathType.DURATION);
 
-    assertAll(
-        () -> assertThat(result.getStations().size()).isEqualTo(3),
-        () -> assertThat(result.getDistance()).isEqualTo(18),
-        () -> assertThat(result.getDuration()).isEqualTo(6),
-        () -> assertThat(result.getFare()).isEqualTo(1_450)
-    );
+    경로_조회_정보_비교(result, 3, 18, 6, 1_450);
   }
 
   @Test
   void 두_역_경로_50KM_초과_최소_시간_경로_조회() {
     PathResponse result = pathService.findPath(남부터미널역, 양재역, PathType.DURATION);
 
+    경로_조회_정보_비교(result, 2, 58, 4, 2_150);
+  }
+
+  private void 경로_조회_정보_비교(PathResponse result, int 지하쳘역_갯수, int 거리, int 시간, int 요금) {
     assertAll(
-        () -> assertThat(result.getStations().size()).isEqualTo(2),
-        () -> assertThat(result.getDistance()).isEqualTo(58),
-        () -> assertThat(result.getDuration()).isEqualTo(4),
-        () -> assertThat(result.getFare()).isEqualTo(2_150)
+        () -> assertThat(result.getStations().size()).isEqualTo(지하쳘역_갯수),
+        () -> assertThat(result.getDistance()).isEqualTo(거리),
+        () -> assertThat(result.getDuration()).isEqualTo(시간),
+        () -> assertThat(result.getFare()).isEqualTo(요금)
     );
   }
